added settings file; moved variables; updated unittest
