+++ ecosystem/settings.py
@@ -1,0 +1,114 @@
+import os
+import platform
+import subprocess
+
+_ON_WINDOWS = (platform.system().lower() == 'windows')
+
+
+def _determine_number_of_cpus():
+    """
+    Number of virtual or physical CPUs on this system, i.e.
+    user/real as output by time(1) when called with an optimally scaling
+    userspace-only program
+    """
+
+    # Python 2.6+
+    try:
+        import multiprocessing
+        return multiprocessing.cpu_count()
+    except (ImportError, NotImplementedError):
+        pass
+
+    # POSIX
+    try:
+        res = int(os.sysconf('SC_NPROCESSORS_ONLN'))
+
+        if res > 0:
+            return res
+    except (AttributeError, ValueError):
+        pass
+
+    # Windows
+    try:
+        res = int(os.environ['NUMBER_OF_PROCESSORS'])
+
+        if res > 0:
+            return res
+    except (KeyError, ValueError):
+        pass
+
+    # jython
+    try:
+        from java.lang import Runtime
+        runtime = Runtime.getRuntime()
+        res = runtime.availableProcessors()
+        if res > 0:
+            return res
+    except ImportError:
+        pass
+
+    # BSD
+    try:
+        sysctl = subprocess.Popen(['sysctl', '-n', 'hw.ncpu'], stdout=subprocess.PIPE)
+        sc_stdout = sysctl.communicate()[0]
+        res = int(sc_stdout)
+
+        if res > 0:
+            return res
+    except (OSError, ValueError):
+        pass
+
+    # Linux
+    try:
+        res = open('/proc/cpuinfo').read().count('processor\t:')
+
+        if res > 0:
+            return res
+    except IOError:
+        pass
+
+    # Solaris
+    try:
+        pseudo_devices = os.listdir('/devices/pseudo/')
+        import re
+        expr = re.compile('^cpuid@[0-9]+$')
+
+        res = 0
+        for pd in pseudo_devices:
+            if expr.match(pd) is not None:
+                res += 1
+
+        if res > 0:
+            return res
+    except OSError:
+        pass
+
+    # Other UNIXes (heuristic)
+    try:
+        try:
+            dmesg = open('/var/run/dmesg.boot').read()
+        except IOError:
+            dmesg_process = subprocess.Popen(['dmesg'], stdout=subprocess.PIPE)
+            dmesg = dmesg_process.communicate()[0]
+
+        res = 0
+        while '\ncpu' + str(res) + ':' in dmesg:
+            res += 1
+
+        if res > 0:
+            return res
+    except OSError:
+        pass
+
+    raise Exception('Can not determine number of CPUs on this system')
+
+
+# set up some global variables
+NUMBER_OF_PROCESSORS = _determine_number_of_cpus()
+MAKE_COMMAND = ['make', '-j', str(NUMBER_OF_PROCESSORS)]
+CLEAN_COMMAND = ['make', 'clean']
+MAKE_TARGET = 'Unix Makefiles'
+if _ON_WINDOWS:
+    MAKE_COMMAND = ['jom']
+    CLEAN_COMMAND = ['jom', 'clean']
+    MAKE_TARGET = 'NMake Makefiles'
